1. 優化錯誤例外 2. 增加超過 Long 最大值例外檢查 3. 調整 buffer 類型

+++ src/main/java/pers/clare/bufferid/controller/BufferFunction.java
@@ -2,5 +2,5 @@ package pers.clare.bufferid.controller;
 
 @FunctionalInterface
 interface BufferFunction<V> {
-    V apply(Integer buffer, String id, String prefix, Integer length);
+    V apply(Long buffer, String id, String prefix, Integer length);
 }
